Tag selector (#141)
Added support to configure a tagSelector closure to select among multiple tags on a commit. Defaults back to selecting the last tag found in the list of tags returned from git, same behavior as before, i.e. when there were no list and the tag for a commit was overwritten.