Adds first try on writing Git repository cache
Due to Gradle execution model this cache is a bit specific:

* has to be able to live for a long time in background and might be run in
  different repo/proejct (Gradle Demon)
* has to be thread safe (parallel builds)

Cache is thread safe via synchronized access - crude but simple and effective.
Sycnhronization cost is neglible in context of this cache, as the exepnsive
operations are single-threaded anyway and if run in parallel all
threads would perform same IO operations on git repo.