GitHub dependabot friendly dependency versions
Looks like GitHub's dependabot is configured for this repo, but it's not picking up a few dependency versions because of unusual Kotlin structures in the build files used to store the version. This PR switches to a more standard / dependabot-friendly way of defining dependency versions.

(Dependabot uses simplistic code parsing to detect dependency versions. It can't handle funky Kotlin stuff)