Switching to jsch implementation by mwiede
Hi! 

I discovered that authorization fails with invalid private key error when private key is generate by a new versions of openssh
(header that is not supported is : -----BEGIN OPENSSH PRIVATE KEY-----)

The problem is that jsch is not actively maintained, but there is a for, which is.

After switching to new implementation, problem is solved. Private key is properly used.