Switch to a recent version of a supported Jsch fork
The version of JSch bundled in the plugin has not been maintained in a
few years so has not kept up-to-date with newer cryptographic keys,
ciphers, and signatures. This causes challenges for users attempting to
use Git with configuration using up-to-date cryptographic
recommendations, as the library only supports SHA1 signatures for RSA
which distributions like OpenSSH have stopped supporting.

The JSch implementation has been switched to an actively maintained fork
and an up-to-date version of BouncyCastle included in the dependencies
to ensure elliptic curves and recent cipher suites are available. The
existing RemoteRejectionTest has been altered to use public key
authentication and an up-to-date container running a recent version of
OpenSSH to allow realistic connectivity testing of remote Git over SSH
connections.

As the Jsch distribution is a multi-version JAR including Java 16 class
files, the plugin has had to be updated to Gradle 7 as the ASM version
used in Gradle 6 is not compatible with Java 16 classes.