implement flag overridenIsClean
This flag is implemented similar to overriddenBranch and allows to set
the "isClean" flag for the repository. Which basically determines if the
working tree is dirty or clean.

One use-case is to speed up the computation when running the plugin
locally as the isClean-calculation is quite time consuming. Also it may
help during development of the version-rules.

This solves one part of issue #736