Expose cached DecoratedVersion, and introduce isSnapshot() to DecoratedVersion
Introduces a method, getDecoratedVersion() in the VersionConfig class, allowing you to easily retrieve the cached decorated version, instead of having to do versionProvider().get(), which looks less nice (imo)

Also introduces a method, isSnapshot() to DecoratedVersion, which allows you to check if the decorated version is a snapshot.
This is passed the value of VersionContact.isSnapshot().
This is convenient for checking if a version is a snapshot or not when building the project.